+++ data/types/common/duration.go
@@ -1,4 +1,4 @@
-package physical
+package common
 
 import (
 	"math"