+++ data/types/common/duration_test.go
@@ -1,4 +1,4 @@
-package physical_test
+package common_test
 
 import (
 	"math"
@@ -8,71 +8,54 @@ import (
 	. "github.com/onsi/gomega"
 
 	dataNormalizer "github.com/tidepool-org/platform/data/normalizer"
-	"github.com/tidepool-org/platform/data/types/activity/physical"
+	dataTypesCommon "github.com/tidepool-org/platform/data/types/common"
+	dataTypesCommonTest "github.com/tidepool-org/platform/data/types/common/test"
 	dataTypesTest "github.com/tidepool-org/platform/data/types/test"
 	errorsTest "github.com/tidepool-org/platform/errors/test"
 	"github.com/tidepool-org/platform/pointer"
 	"github.com/tidepool-org/platform/structure"
 	structureValidator "github.com/tidepool-org/platform/structure/validator"
-	"github.com/tidepool-org/platform/test"
 )
-
-func NewDuration() *physical.Duration {
-	datum := physical.NewDuration()
-	datum.Units = pointer.FromString(test.RandomStringFromArray(physical.DurationUnits()))
-	datum.Value = pointer.FromFloat64(test.RandomFloat64FromRange(physical.DurationValueRangeForUnits(datum.Units)))
-	return datum
-}
-
-func CloneDuration(datum *physical.Duration) *physical.Duration {
-	if datum == nil {
-		return nil
-	}
-	clone := physical.NewDuration()
-	clone.Units = pointer.CloneString(datum.Units)
-	clone.Value = pointer.CloneFloat64(datum.Value)
-	return clone
-}
 
 var _ = Describe("Duration", func() {
 	It("DurationUnitsHours is expected", func() {
-		Expect(physical.DurationUnitsHours).To(Equal("hours"))
+		Expect(dataTypesCommon.DurationUnitsHours).To(Equal("hours"))
 	})
 
 	It("DurationUnitsMinutes is expected", func() {
-		Expect(physical.DurationUnitsMinutes).To(Equal("minutes"))
+		Expect(dataTypesCommon.DurationUnitsMinutes).To(Equal("minutes"))
 	})
 
 	It("DurationUnitsSeconds is expected", func() {
-		Expect(physical.DurationUnitsSeconds).To(Equal("seconds"))
+		Expect(dataTypesCommon.DurationUnitsSeconds).To(Equal("seconds"))
 	})
 
 	It("DurationValueHoursMaximum is expected", func() {
-		Expect(physical.DurationValueHoursMaximum).To(Equal(168.0))
+		Expect(dataTypesCommon.DurationValueHoursMaximum).To(Equal(168.0))
 	})
 
 	It("DurationValueHoursMinimum is expected", func() {
-		Expect(physical.DurationValueHoursMinimum).To(Equal(0.0))
+		Expect(dataTypesCommon.DurationValueHoursMinimum).To(Equal(0.0))
 	})
 
 	It("DurationValueMinutesMaximum is expected", func() {
-		Expect(physical.DurationValueMinutesMaximum).To(Equal(10080.0))
+		Expect(dataTypesCommon.DurationValueMinutesMaximum).To(Equal(10080.0))
 	})
 
 	It("DurationValueMinutesMinimum is expected", func() {
-		Expect(physical.DurationValueMinutesMinimum).To(Equal(0.0))
+		Expect(dataTypesCommon.DurationValueMinutesMinimum).To(Equal(0.0))
 	})
 
 	It("DurationValueSecondsMaximum is expected", func() {
-		Expect(physical.DurationValueSecondsMaximum).To(Equal(604800.0))
+		Expect(dataTypesCommon.DurationValueSecondsMaximum).To(Equal(604800.0))
 	})
 
 	It("DurationValueSecondsMinimum is expected", func() {
-		Expect(physical.DurationValueSecondsMinimum).To(Equal(0.0))
+		Expect(dataTypesCommon.DurationValueSecondsMinimum).To(Equal(0.0))
 	})
 
 	It("DurationUnits returns expected", func() {
-		Expect(physical.DurationUnits()).To(Equal([]string{"hours", "minutes", "seconds"}))
+		Expect(dataTypesCommon.DurationUnits()).To(Equal([]string{"hours", "minutes", "seconds"}))
 	})
 
 	Context("ParseDuration", func() {
@@ -81,7 +64,7 @@ var _ = Describe("Duration", func() {
 
 	Context("NewDuration", func() {
 		It("returns the expected datum", func() {
-			Expect(physical.NewDuration()).To(Equal(&physical.Duration{}))
+			Expect(dataTypesCommon.NewDuration()).To(Equal(&dataTypesCommon.Duration{}))
 		})
 	})
 
@@ -92,16 +75,16 @@ var _ = Describe("Duration", func() {
 
 		Context("Validate", func() {
 			DescribeTable("validates the datum",
-				func(mutator func(datum *physical.Duration), expectedErrors ...error) {
-					datum := NewDuration()
+				func(mutator func(datum *dataTypesCommon.Duration), expectedErrors ...error) {
+					datum := dataTypesCommonTest.NewDuration()
 					mutator(datum)
 					dataTypesTest.ValidateWithExpectedOrigins(datum, structure.Origins(), expectedErrors...)
 				},
 				Entry("succeeds",
-					func(datum *physical.Duration) {},
+					func(datum *dataTypesCommon.Duration) {},
 				),
 				Entry("units missing; value missing",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = nil
 						datum.Value = nil
 					},
@@ -109,35 +92,35 @@ var _ = Describe("Duration", func() {
 					errorsTest.WithPointerSource(structureValidator.ErrorValueNotExists(), "/value"),
 				),
 				Entry("units missing; value out of range (lower)",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = nil
 						datum.Value = pointer.FromFloat64(-0.1)
 					},
 					errorsTest.WithPointerSource(structureValidator.ErrorValueNotExists(), "/units"),
 				),
 				Entry("units missing; value in range (lower)",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = nil
 						datum.Value = pointer.FromFloat64(0.0)
 					},
 					errorsTest.WithPointerSource(structureValidator.ErrorValueNotExists(), "/units"),
 				),
 				Entry("units missing; value in range (upper)",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = nil
 						datum.Value = pointer.FromFloat64(604800.0)
 					},
 					errorsTest.WithPointerSource(structureValidator.ErrorValueNotExists(), "/units"),
 				),
 				Entry("units missing; value out of range (upper)",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = nil
 						datum.Value = pointer.FromFloat64(604800.1)
 					},
 					errorsTest.WithPointerSource(structureValidator.ErrorValueNotExists(), "/units"),
 				),
 				Entry("units invalid; value missing",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = pointer.FromString("invalid")
 						datum.Value = nil
 					},
@@ -145,134 +128,134 @@ var _ = Describe("Duration", func() {
 					errorsTest.WithPointerSource(structureValidator.ErrorValueNotExists(), "/value"),
 				),
 				Entry("units invalid; value out of range (lower)",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = pointer.FromString("invalid")
 						datum.Value = pointer.FromFloat64(-0.1)
 					},
 					errorsTest.WithPointerSource(structureValidator.ErrorValueStringNotOneOf("invalid", []string{"hours", "minutes", "seconds"}), "/units"),
 				),
 				Entry("units invalid; value in range (lower)",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = pointer.FromString("invalid")
 						datum.Value = pointer.FromFloat64(0.0)
 					},
 					errorsTest.WithPointerSource(structureValidator.ErrorValueStringNotOneOf("invalid", []string{"hours", "minutes", "seconds"}), "/units"),
 				),
 				Entry("units invalid; value in range (upper)",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = pointer.FromString("invalid")
 						datum.Value = pointer.FromFloat64(604800.0)
 					},
 					errorsTest.WithPointerSource(structureValidator.ErrorValueStringNotOneOf("invalid", []string{"hours", "minutes", "seconds"}), "/units"),
 				),
 				Entry("units invalid; value out of range (upper)",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = pointer.FromString("invalid")
 						datum.Value = pointer.FromFloat64(604800.1)
 					},
 					errorsTest.WithPointerSource(structureValidator.ErrorValueStringNotOneOf("invalid", []string{"hours", "minutes", "seconds"}), "/units"),
 				),
 				Entry("units hours; value missing",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = pointer.FromString("hours")
 						datum.Value = nil
 					},
 					errorsTest.WithPointerSource(structureValidator.ErrorValueNotExists(), "/value"),
 				),
 				Entry("units hours; value out of range (lower)",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = pointer.FromString("hours")
 						datum.Value = pointer.FromFloat64(-0.1)
 					},
 					errorsTest.WithPointerSource(structureValidator.ErrorValueNotInRange(-0.1, 0.0, 168.0), "/value"),
 				),
 				Entry("units hours; value in range (lower)",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = pointer.FromString("hours")
 						datum.Value = pointer.FromFloat64(0.0)
 					},
 				),
 				Entry("units hours; value in range (upper)",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = pointer.FromString("hours")
 						datum.Value = pointer.FromFloat64(168.0)
 					},
 				),
 				Entry("units hours; value out of range (upper)",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = pointer.FromString("hours")
 						datum.Value = pointer.FromFloat64(168.1)
 					},
 					errorsTest.WithPointerSource(structureValidator.ErrorValueNotInRange(168.1, 0.0, 168.0), "/value"),
 				),
 				Entry("units minutes; value missing",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = pointer.FromString("minutes")
 						datum.Value = nil
 					},
 					errorsTest.WithPointerSource(structureValidator.ErrorValueNotExists(), "/value"),
 				),
 				Entry("units minutes; value out of range (lower)",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = pointer.FromString("minutes")
 						datum.Value = pointer.FromFloat64(-0.1)
 					},
 					errorsTest.WithPointerSource(structureValidator.ErrorValueNotInRange(-0.1, 0.0, 10080.0), "/value"),
 				),
 				Entry("units minutes; value in range (lower)",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = pointer.FromString("minutes")
 						datum.Value = pointer.FromFloat64(0.0)
 					},
 				),
 				Entry("units minutes; value in range (upper)",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = pointer.FromString("minutes")
 						datum.Value = pointer.FromFloat64(10080.0)
 					},
 				),
 				Entry("units minutes; value out of range (upper)",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = pointer.FromString("minutes")
 						datum.Value = pointer.FromFloat64(10080.1)
 					},
 					errorsTest.WithPointerSource(structureValidator.ErrorValueNotInRange(10080.1, 0.0, 10080.0), "/value"),
 				),
 				Entry("units seconds; value missing",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = pointer.FromString("seconds")
 						datum.Value = nil
 					},
 					errorsTest.WithPointerSource(structureValidator.ErrorValueNotExists(), "/value"),
 				),
 				Entry("units seconds; value out of range (lower)",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = pointer.FromString("seconds")
 						datum.Value = pointer.FromFloat64(-0.1)
 					},
 					errorsTest.WithPointerSource(structureValidator.ErrorValueNotInRange(-0.1, 0.0, 604800.0), "/value"),
 				),
 				Entry("units seconds; value in range (lower)",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = pointer.FromString("seconds")
 						datum.Value = pointer.FromFloat64(0.0)
 					},
 				),
 				Entry("units seconds; value in range (upper)",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = pointer.FromString("seconds")
 						datum.Value = pointer.FromFloat64(604800.0)
 					},
 				),
 				Entry("units seconds; value out of range (upper)",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = pointer.FromString("seconds")
 						datum.Value = pointer.FromFloat64(604800.1)
 					},
 					errorsTest.WithPointerSource(structureValidator.ErrorValueNotInRange(604800.1, 0.0, 604800.0), "/value"),
 				),
 				Entry("multiple errors",
-					func(datum *physical.Duration) {
+					func(datum *dataTypesCommon.Duration) {
 						datum.Units = nil
 						datum.Value = nil
 					},
@@ -284,11 +267,11 @@ var _ = Describe("Duration", func() {
 
 		Context("Normalize", func() {
 			DescribeTable("normalizes the datum",
-				func(mutator func(datum *physical.Duration)) {
+				func(mutator func(datum *dataTypesCommon.Duration)) {
 					for _, origin := range structure.Origins() {
-						datum := NewDuration()
+						datum := dataTypesCommonTest.NewDuration()
 						mutator(datum)
-						expectedDatum := CloneDuration(datum)
+						expectedDatum := dataTypesCommonTest.CloneDuration(datum)
 						normalizer := dataNormalizer.New()
 						Expect(normalizer).ToNot(BeNil())
 						datum.Normalize(normalizer.WithOrigin(origin))
@@ -298,22 +281,22 @@ var _ = Describe("Duration", func() {
 					}
 				},
 				Entry("does not modify the datum",
-					func(datum *physical.Duration) {},
+					func(datum *dataTypesCommon.Duration) {},
 				),
 				Entry("does not modify the datum; units missing",
-					func(datum *physical.Duration) { datum.Units = nil },
+					func(datum *dataTypesCommon.Duration) { datum.Units = nil },
 				),
 				Entry("does not modify the datum; units hours",
-					func(datum *physical.Duration) { datum.Units = pointer.FromString("hours") },
+					func(datum *dataTypesCommon.Duration) { datum.Units = pointer.FromString("hours") },
 				),
 				Entry("does not modify the datum; units minutes",
-					func(datum *physical.Duration) { datum.Units = pointer.FromString("minutes") },
+					func(datum *dataTypesCommon.Duration) { datum.Units = pointer.FromString("minutes") },
 				),
 				Entry("does not modify the datum; units seconds",
-					func(datum *physical.Duration) { datum.Units = pointer.FromString("seconds") },
+					func(datum *dataTypesCommon.Duration) { datum.Units = pointer.FromString("seconds") },
 				),
 				Entry("does not modify the datum; value missing",
-					func(datum *physical.Duration) { datum.Value = nil },
+					func(datum *dataTypesCommon.Duration) { datum.Value = nil },
 				),
 			)
 		})
@@ -321,31 +304,31 @@ var _ = Describe("Duration", func() {
 
 	Context("DurationValueRangeForUnits", func() {
 		It("returns expected range for units missing", func() {
-			minimum, maximum := physical.DurationValueRangeForUnits(nil)
+			minimum, maximum := dataTypesCommon.DurationValueRangeForUnits(nil)
 			Expect(minimum).To(Equal(-math.MaxFloat64))
 			Expect(maximum).To(Equal(math.MaxFloat64))
 		})
 
 		It("returns expected range for units invalid", func() {
-			minimum, maximum := physical.DurationValueRangeForUnits(pointer.FromString("invalid"))
+			minimum, maximum := dataTypesCommon.DurationValueRangeForUnits(pointer.FromString("invalid"))
 			Expect(minimum).To(Equal(-math.MaxFloat64))
 			Expect(maximum).To(Equal(math.MaxFloat64))
 		})
 
 		It("returns expected range for units hours", func() {
-			minimum, maximum := physical.DurationValueRangeForUnits(pointer.FromString("hours"))
+			minimum, maximum := dataTypesCommon.DurationValueRangeForUnits(pointer.FromString("hours"))
 			Expect(minimum).To(Equal(0.0))
 			Expect(maximum).To(Equal(168.0))
 		})
 
 		It("returns expected range for units minutes", func() {
-			minimum, maximum := physical.DurationValueRangeForUnits(pointer.FromString("minutes"))
+			minimum, maximum := dataTypesCommon.DurationValueRangeForUnits(pointer.FromString("minutes"))
 			Expect(minimum).To(Equal(0.0))
 			Expect(maximum).To(Equal(10080.0))
 		})
 
 		It("returns expected range for units seconds", func() {
-			minimum, maximum := physical.DurationValueRangeForUnits(pointer.FromString("seconds"))
+			minimum, maximum := dataTypesCommon.DurationValueRangeForUnits(pointer.FromString("seconds"))
 			Expect(minimum).To(Equal(0.0))
 			Expect(maximum).To(Equal(604800.0))
 		})